requirements.txt
torchxrayvision
numpy
pillow
pydicom
scikit-image
matplotlib

